update: Updated the valorant slash command to be under one command using sub commands
